Initial image hashing, accumulate keywords
Image files (.png, .jpg) hash only the image data, so two files with the
same image, but different metadata, will be considered the same picture.

When archiving files, keywords are accumulated, i.e. keywords from the
duplicate file being archived will be added to the existing image in the
archive.